we-167 Add support doc algolia indexing and search input
